+++ app/controllers/rails_pallet/application_controller.rb
@@ -1,0 +1,4 @@
+module RailsPallet
+  class ApplicationController < ActionController::Base
+  end
+end