+++ db/migrate/20150612152328_create_paperclip_upload_uploads.rb
@@ -1,7 +1,0 @@
-class CreatePaperclipUploadUploads < ActiveRecord::Migration
-  def change
-    create_table :paperclip_upload_uploads do |t|
-      t.timestamps null: false
-    end
-  end
-end