+++ spec/factories/paperclip_upload/uploads.rb
@@ -1,7 +1,0 @@
-include ActionDispatch::TestProcess
-
-FactoryGirl.define do
-  factory :upload, :class => 'PaperclipUpload::Upload' do
-    file { fixture_file_upload(Rails.root.join('spec', 'assets', 'bukowski.jpg'), 'image/png') }
-  end
-end